requirements.txt
requests>=2.31.0
textual>=0.41.0
openai>=1.0.0
python-dotenv>=1.0.0